Expose the headless kafka service to allow CLI
The Kafka CLI tools are very useful for debugging, but also for
administrating a Kafka cluster. Unfortunately, the Helm chart currently
installs Kafka in such a way that the charts don't work. This is because for
some operations it needs the headless service to be accessible. This commit
removes the hardcoded non-exposure of the headless service to allow the CLI
to work as designed